[stream] rename dns syncing package to legacysync
