+++ api/service/legacysync/downloader/client.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	pb "github.com/harmony-one/harmony/api/service/syncing/downloader/proto"
+	pb "github.com/harmony-one/harmony/api/service/legacysync/downloader/proto"
 	"github.com/harmony-one/harmony/internal/utils"
 	"google.golang.org/grpc"
 )
