+++ api/service/legacysync/downloader/gen.sh
