+++ api/service/legacysync/downloader/server.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"net"
 
-	pb "github.com/harmony-one/harmony/api/service/syncing/downloader/proto"
+	pb "github.com/harmony-one/harmony/api/service/legacysync/downloader/proto"
 	"github.com/harmony-one/harmony/internal/utils"
 
 	"google.golang.org/grpc"
