+++ api/service/legacysync/errors.go
@@ -1,4 +1,4 @@
-package syncing
+package legacysync
 
 import "errors"
